+++ apps/full-site-editing/full-site-editing-plugin/starter-page-templates/page-template-modal/components/template-selector-preview.js
@@ -2,20 +2,24 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import { isEmpty, debounce } from 'lodash';
+/* eslint-disable import/no-extraneous-dependencies */
+import { isEmpty, isArray, debounce } from 'lodash';
+/* eslint-enable import/no-extraneous-dependencies */
 
 /**
  * WordPress dependencies
  */
+/* eslint-disable import/no-extraneous-dependencies */
 import { __ } from '@wordpress/i18n';
-import { BlockPreview } from '@wordpress/block-editor';
 import { Disabled } from '@wordpress/components';
 import { useState, useEffect, useLayoutEffect, useRef, useReducer } from '@wordpress/element';
+/* eslint-enable import/no-extraneous-dependencies */
 
 /**
  * Internal dependencies
  */
 import PreviewTemplateTitle from './preview-template-title';
+import BlockPreview from './block-preview';
 
 const TemplateSelectorPreview = ( { blocks, viewportWidth, title } ) => {
 	const THRESHOLD_RESIZE = 300;
@@ -42,7 +46,9 @@ const TemplateSelectorPreview = ( { blocks, viewportWidth, title } ) => {
 			}
 
 			// Try to get the preview content element.
-			const previewContainerEl = ref.current.querySelector( '.block-editor-block-preview__content' );
+			const previewContainerEl = ref.current.querySelector(
+				'.block-editor-block-preview__content'
+			);
 			if ( ! previewContainerEl ) {
 				return;
 			}
@@ -63,7 +69,7 @@ const TemplateSelectorPreview = ( { blocks, viewportWidth, title } ) => {
 	}, [ blocks ] );
 
 	useEffect( () => {
-		if ( ! blocks.length ) {
+		if ( ! blocks || ! blocks.length ) {
 			return;
 		}
 
@@ -80,7 +86,7 @@ const TemplateSelectorPreview = ( { blocks, viewportWidth, title } ) => {
 		};
 	}, [ blocks ] );
 
-	if ( isEmpty( blocks ) ) {
+	if ( isEmpty( blocks ) || ! isArray( blocks ) ) {
 		return (
 			<div className={ previewElClasses }>
 				<div className="template-selector-preview__placeholder">