+++ client/blocks/app-promo/index.jsx
@@ -17,7 +17,7 @@ import Gridicon from 'components/gridicon';
 import { localize } from 'i18n-calypso';
 import { recordTracksEvent } from 'state/analytics/actions';
 import wpcom from 'lib/wp';
-import Dialog from 'components/dialog'
+import Dialog from 'components/dialog';
 import { fetchUserSettings } from 'state/user-settings/actions';
 import getUserSettings from 'state/selectors/get-user-settings';
 
@@ -106,15 +106,15 @@ export class AppPromo extends React.Component {
 	};
 
 	sendMagicLink = () => {
-		this.recordClickEvent()
+		this.recordClickEvent();
 
 		const email = this.props.userSettings.user_email;
 		wpcom.undocumented().requestMagicLoginEmail( {
 			email,
-			'infer': true,
-			'scheme': 'wordpress',
-		} )
-		
+			infer: true,
+			scheme: 'wordpress',
+		} );
+
 		this.onShowDialog();
 
 		return false;
@@ -127,8 +127,8 @@ export class AppPromo extends React.Component {
 	onCloseDialog = () => {
 		this.setState( { showDialog: false } );
 	};
-	
-	desktopPromo = ( promoItem ) => {
+
+	desktopPromo = promoItem => {
 		const { location, translate } = this.props;
 
 		return (
@@ -159,14 +159,12 @@ export class AppPromo extends React.Component {
 					{ promoItem.message }
 				</a>
 			</div>
-		)
-	}
-	
+		);
+	};
+
 	mobilePromo = () => {
 		const { translate } = this.props;
-		const buttons = [
-			{ action: 'cancel', label: translate( 'OK' ) },
-		];
+		const buttons = [ { action: 'cancel', label: translate( 'OK' ) } ];
 
 		return (
 			<div className="app-promo">
@@ -192,20 +190,29 @@ export class AppPromo extends React.Component {
 					/>
 					{ 'WordPress.com in the palm of your hands — download the mobile app.' }
 				</button>
-				<Dialog className="app-promo__dialog" isVisible={ this.state.showDialog } buttons={ buttons } onClose={ this.onCloseDialog }>
+				<Dialog
+					className="app-promo__dialog"
+					isVisible={ this.state.showDialog }
+					buttons={ buttons }
+					onClose={ this.onCloseDialog }
+				>
 					<h1>{ translate( 'Check your mail!' ) }</h1>
-					<p>{ translate( "We've sent you an email with links to download and effortlessly log in to the mobile app. Be sure to use them from your mobile device!" ) }</p>
+					<p>
+						{ translate(
+							"We've sent you an email with links to download and effortlessly log in to the mobile app. Be sure to use them from your mobile device!"
+						) }
+					</p>
 				</Dialog>
 			</div>
-		)
-	}
+		);
+	};
 
 	render() {
 		if ( ! this.state.showPromo ) {
 			return null;
 		}
 		const { promoItem } = this.state;
-		return ( promoItem.type === 'mobile' ) ? this.mobilePromo() : this.desktopPromo( promoItem );
+		return promoItem.type === 'mobile' ? this.mobilePromo() : this.desktopPromo( promoItem );
 	}
 }