+++ client/blocks/image-editor/image-editor-toolbar.jsx
@@ -117,7 +117,9 @@ export class ImageEditorToolbar extends Component {
 		const items = [
 			{
 				action: AspectRatios.FREE,
-				label: translate( 'Freeform', {context: 'Option in image editor used to crop images using freeform aspect ratio'} ),
+				label: translate( 'Freeform', {
+					context: 'Option in image editor used to crop images using freeform aspect ratio',
+				} ),
 			},
 			{
 				action: AspectRatios.ORIGINAL,