+++ client/components/popover/index.jsx
@@ -10,7 +10,7 @@ import { connect } from 'react-redux';
 import debugFactory from 'debug';
 import classNames from 'classnames';
 import clickOutside from 'click-outside';
-import { uniqueId } from 'lodash';
+import { defer, uniqueId } from 'lodash';
 
 /**
  * Internal dependencies
@@ -45,7 +45,6 @@ class Popover extends Component {
 		closeOnEsc: PropTypes.bool,
 		id: PropTypes.string,
 		ignoreContext: PropTypes.shape( { getDOMNode: PropTypes.function } ),
-		isFocusOnShow: PropTypes.bool,
 		isRtl: PropTypes.bool,
 		isVisible: PropTypes.bool,
 		position: PropTypes.oneOf( [
@@ -58,7 +57,6 @@ class Popover extends Component {
 			'left',
 			'top left',
 		] ),
-		rootClassName: PropTypes.string,
 		showDelay: PropTypes.number,
 		onClose: PropTypes.func,
 		onShow: PropTypes.func,
@@ -76,7 +74,6 @@ class Popover extends Component {
 		autoRtl: true,
 		className: '',
 		closeOnEsc: true,
-		isFocusOnShow: false,
 		isRtl: false,
 		isVisible: false,
 		position: 'top',
@@ -284,7 +281,7 @@ class Popover extends Component {
 	}
 
 	focusPopover() {
-		if ( ! this.props.isFocusOnShow || ! this.popoverNode ) {
+		if ( ! this.popoverNode ) {
 			return null;
 		}
 
@@ -311,7 +308,7 @@ class Popover extends Component {
 
 		this.setPosition();
 
-		this.focusPopover();
+		defer( () => this.focusPopover() );
 	}
 
 	getPositionClass( position = this.props.position ) {
@@ -502,14 +499,15 @@ class Popover extends Component {
 		this.debug( 'rendering ...' );
 
 		return (
-			<RootChild
-				aria-label={ this.props[ 'aria-label' ] }
-				className={ this.props.rootClassName }
-				id={ this.id }
-				role="tooltip"
-				tabIndex={ this.props.isFocusOnShow ? 0 : null }
-			>
-				<div style={ this.getStylePosition() } className={ classes }>
+			<RootChild>
+				<div
+					aria-label={ this.props[ 'aria-label' ] }
+					id={ this.id }
+					role="tooltip"
+					tabIndex="-1"
+					style={ this.getStylePosition() }
+					className={ classes }
+				>
 					<div className="popover__arrow" />
 					<div ref={ this.setDOMBehavior } className="popover__inner">
 						{ this.props.children }