+++ client/components/popover/menu.jsx
@@ -21,7 +21,6 @@ class PopoverMenu extends Component {
 		onClose: PropTypes.func.isRequired,
 		position: PropTypes.string,
 		className: PropTypes.string,
-		rootClassName: PropTypes.string,
 		popoverComponent: PropTypes.func,
 		popoverTitle: PropTypes.string, // used by ReaderPopover
 		customPosition: PropTypes.object,
@@ -50,7 +49,6 @@ class PopoverMenu extends Component {
 			isVisible,
 			popoverTitle,
 			position,
-			rootClassName,
 		} = this.props;
 
 		return (
@@ -64,7 +62,6 @@ class PopoverMenu extends Component {
 				isVisible={ isVisible }
 				popoverTitle={ popoverTitle }
 				position={ position }
-				rootClassName={ rootClassName }
 			>
 				<div
 					ref={ this.menu }