+++ client/components/suggestion-search/index.jsx
@@ -6,12 +6,12 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { escapeRegExp, noop } from 'lodash';
-import Gridicon from 'components/gridicon';
 
 /**
  * Internal dependencies
  */
 import FormTextInput from 'components/forms/form-text-input';
+import Gridicon from 'components/gridicon';
 import Suggestions from 'components/suggestions';
 import Spinner from 'components/spinner';
 
@@ -26,7 +26,12 @@ class SuggestionSearch extends Component {
 		placeholder: PropTypes.string,
 		onChange: PropTypes.func,
 		onSelect: PropTypes.func,
-		suggestions: PropTypes.array,
+		suggestions: PropTypes.arrayOf(
+			PropTypes.shape( {
+				label: PropTypes.string.isRequired,
+				category: PropTypes.string,
+			} )
+		),
 		value: PropTypes.string,
 		autoFocus: PropTypes.bool,
 		railcar: PropTypes.object,
@@ -87,7 +92,7 @@ class SuggestionSearch extends Component {
 			return [];
 		}
 
-		return this.props.suggestions.map( hint => ( { label: hint } ) );
+		return this.props.suggestions;
 	}
 
 	getSuggestionLabel( suggestionPosition ) {