+++ client/layout/logged-out.jsx
@@ -1,9 +1,6 @@
-/** @format */
-
 /**
  * External dependencies
  */
-
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
@@ -26,6 +23,11 @@ import { getSection, masterbarIsVisible } from 'state/ui/selectors';
 import BodySectionCssClass from './body-section-css-class';
 import GdprBanner from 'blocks/gdpr-banner';
 
+/**
+ * Style dependencies
+ */
+import './style.scss';
+
 // Returns true if given section should display sidebar for logged out users.
 const hasSidebar = section => {
 	if ( section.name === 'devdocs' ) {
@@ -39,6 +41,7 @@ const hasSidebar = section => {
 const LayoutLoggedOut = ( {
 	currentRoute,
 	isJetpackLogin,
+	isPopup,
 	isJetpackWooCommerceFlow,
 	wccomFrom,
 	masterbarIsHidden,
@@ -60,6 +63,7 @@ const LayoutLoggedOut = ( {
 		'has-no-sidebar': ! hasSidebar( section ),
 		'has-no-masterbar': masterbarIsHidden,
 		'is-jetpack-login': isJetpackLogin,
+		'is-popup': isPopup,
 		'is-jetpack-woocommerce-flow':
 			config.isEnabled( 'jetpack/connect/woocommerce' ) && isJetpackWooCommerceFlow,
 		'is-wccom-oauth-flow':
@@ -134,6 +138,7 @@ export default connect( state => {
 	const currentRoute = getCurrentRoute( state );
 	const isJetpackLogin = startsWith( currentRoute, '/log-in/jetpack' );
 	const noMasterbarForRoute = startsWith( currentRoute, '/log-in/jetpack' );
+	const isPopup = '1' === get( getCurrentQueryArguments( state ), 'is_popup' );
 	const noMasterbarForSection = 'signup' === section.name || 'jetpack-connect' === section.name;
 	const isJetpackWooCommerceFlow =
 		'woocommerce-setup-wizard' === get( getCurrentQueryArguments( state ), 'from' );
@@ -142,6 +147,7 @@ export default connect( state => {
 	return {
 		currentRoute,
 		isJetpackLogin,
+		isPopup,
 		isJetpackWooCommerceFlow,
 		wccomFrom,
 		masterbarIsHidden: