+++ client/lib/route/index.ts
@@ -17,7 +17,7 @@ const appendQueryString = ( basepath: string, querystring: string ): string =>
 	basepath + ( querystring ? '?' + querystring : '' );
 
 interface QueryArgs {
-	[key: string]: Primitive;
+	[ key: string ]: Primitive;
 }
 export const addQueryArgs = ( args: QueryArgs, url: URL ): URL => {
 	if ( 'object' !== typeof args ) {