+++ client/my-sites/checkout/checkout/test/web-payment-box.js
@@ -11,7 +11,6 @@ import { shallow } from 'enzyme';
 import { identity } from 'lodash';
 
 import { WebPaymentBox } from '../web-payment-box';
-import { BEFORE_SUBMIT } from 'lib/store-transactions/step-types';
 import PaymentCountrySelect from 'components/payment-country-select';
 import { setTaxCountryCode, setTaxPostalCode } from 'lib/upgrades/actions/cart';
 
@@ -49,11 +48,10 @@ describe( 'WebPaymentBox', () => {
 
 	const defaultProps = {
 		cart: defaultCart,
+		disablePostalCodeDebounce: true,
 		translate: identity,
 		countriesList: [ 'TEST_COUNTRY_CODE' ],
 		onSubmit: jest.fn(),
-		transactionStep: { name: BEFORE_SUBMIT },
-		transaction: {},
 	};
 
 	const context = {