+++ client/state/data-layer/wpcom/concierge/schedules/appointments/cancel/test/index.js
@@ -28,9 +28,7 @@ describe( 'wpcom-api', () => {
 				http(
 					{
 						method: 'POST',
-						path: `/concierge/schedules/${ action.scheduleId }/appointments/${
-							action.appointmentId
-						}/cancel`,
+						path: `/concierge/schedules/${ action.scheduleId }/appointments/${ action.appointmentId }/cancel`,
 						apiNamespace: 'wpcom/v2',
 						body: {},
 					},