+++ packages/calypso-build/CHANGELOG.md
@@ -1,8 +1,13 @@
-# _unreleased_
+# 4.2.0
+- Support CommonJS/ESM compilation by adding a `modules` option (and `MODULES` env variable) to the `babel/default` preset.
 
-- Add config options to file-loader
+# 4.1.0
+
+- Add config options to file-loader.
 - Add `enzyme-to-json` serializer to `jest-preset.js`.
 - Make Jest ignore transpiled files in `dist/`.
+- Handle `TypeScript` files in `transpile` command.
+- Use `require.resolve` to resolve babel plugins and presets.
 
 # 4.0.1